requirements.txt
openai
python-dotenv
faiss-cpu
transformers
torch==2.1.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
tiktoken
numpy==1.26.4
pinecone-client